+++ python/pyfury/__init__.py
@@ -32,6 +32,7 @@ from pyfury._fury import (  # noqa: F401,F403,F811 # pylint: disable=unused-impo
     Language,
     OpaqueObject,
 )
+
 if ENABLE_FURY_CYTHON_SERIALIZATION:
     from pyfury._serialization import (  # noqa: F401,F811
         Fury,
